requirements.txt
# Language servers для Zed
pyright>=1.1.408
ruff>=0.3.0

# Основные зависимости проекта
dagster>=1.6.0
dagster-webserver>=0.24.0
dagster-postgres>=0.22.0
dagster-duckdb>=0.22.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0
polars>=0.20.0
duckdb>=0.10.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
minio>=7.2.0
pyarrow>=14.0.0
connectorx>=0.3.2

# Инструменты разработки
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
ipython>=8.20.0
pandas>=2.0.0
